End of Day Commit
End of Day Commit. Last thing we need to work on is testing, and output. I have created a show map method in the game engine, Julia, just connect it with your dungeon methods to print out a map and we should be good. I will work on the status. Theoretically this code should run if you run main()